color masks

+++ vis/make_color_masks.py
@@ -13,13 +13,17 @@ vis = ImageVisualizer(project)
 
 for im in im_list:
     print im
+
+    output_path = os.path.join(output_dir, im)
+    if os.path.exists(output_path):
+        continue
+
+    if not os.path.exists(os.path.dirname(output_path)):
+        os.makedirs(os.path.dirname(output_path))
+
     cm, cm_path = vis.get_category_mask(im)
     if cm is not None:
         cm_color, cm_color_path = vis.add_color(cm)
-        output_path = os.path.join(output_dir, im)
-
-        if not os.path.exists(os.path.dirname(output_path)):
-            os.makedirs(os.path.dirname(output_path))
         os.rename(cm_color_path, output_path)
     else:
         print "Skipping", im
